requirements.txt
Package              Version
-------------------- ---------
asgiref              3.5.2
certifi              2022.9.24
charset-normalizer   2.1.1
Django               4.1.1
django-debug-toolbar 3.7.0
djangorestframework  3.14.0
idna                 3.4
pip                  22.0.2
pytz                 2022.4
requests             2.28.1
setuptools           59.6.0
sqlparse             0.4.2
urllib3              1.26.12
